requirements.txt
selenium
pandas
schedule
python-dotenv
webdriver-manager
openpyxl
customtkinter
matplotlib
packaging
pillow
pyinstaller
